requirements.txt
PyYAML
loki-logger-handler
loguru
requests